% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel.R
\name{babel}
\alias{babel}
\alias{babel_file}
\alias{babel_options}
\title{Babel}
\usage{
babel_file(input, options = babel_options(), output = NULL)

babel(code, options = babel_options())

babel_options(presets = list("es2015"), sourceType = "script", ...)
}
\arguments{
\item{input}{Path to JavaScript file.}

\item{options}{Options for terser, see \url{https://babeljs.io/docs/en/options}.}

\item{output}{Path where to write transformed code.}

\item{code}{Character vector where each element represent a line of CSS code.}

\item{presets}{Preset to use, default to \code{"es2015"}.}

\item{sourceType}{Default to \code{"script"} : parse the file using the ECMAScript Script grammar.}

\item{...}{Other options to use, see \url{https://babeljs.io/docs/en/options} for details.}
}
\value{
a \code{list}.

a \code{list}.
}
\description{
Babel is a tool that helps you write code in the latest version of
 JavaScript. When your supported environments don't support certain features
 natively, Babel will help you compile those features down to a supported version.
}
\examples{
# Template literals are not supported in ES5
babel("var filename = `${Date.now()}.png`;")

# Arrow function neither
babel("[1, 2, 3].map((n) => n + 1);")

}
