% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crass.R
\name{crass}
\alias{crass}
\alias{crass_file}
\title{Crass - CSS minification and pretty printing}
\usage{
crass_file(input, pretty = FALSE, output = NULL)

crass(code, pretty = FALSE)
}
\arguments{
\item{input}{Path to a CSS file.}

\item{pretty}{Pretty print result? Otherwise result is minified.}

\item{output}{Path where to write optimized code.}

\item{code}{Character vector where each element represent a line of CSS code.}
}
\value{
a \code{character}.
}
\description{
Crass is one of only a handful of CSS minifiers that
 creates a full parse tree of the CSS. Most other CSS minifiers
 operate on the string source instead, which makes it impossible
 to perform all types of optimizations.
}
\examples{

crass("b {font-weight: bold;}")

path <- system.file("testfiles/example.css", package = "jstools")
crass_file(input = path)
}
