% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csso.R
\name{csso}
\alias{csso}
\alias{csso_file}
\alias{csso_options}
\title{CSSO - CSS optimizer}
\usage{
csso_file(input, options = csso_options(), output = NULL)

csso(code, options = csso_options())

csso_options(restructure = TRUE, ...)
}
\arguments{
\item{input}{Path to one or more CSS files.}

\item{options}{Options for CSSO, see \url{https://github.com/css/csso}.}

\item{output}{Path where to write minified code.}

\item{code}{Character vector where each element represent a line of CSS code.}

\item{restructure}{Restructure code or not.}

\item{...}{Other options to use, see \url{https://github.com/css/csso} for details.}
}
\value{
a \code{character}.
}
\description{
CSSO (CSS Optimizer) is a CSS minifier. It performs three sort
 of transformations: cleaning (removing redundant), compression (replacement
 for shorter form) and restructuring (merge of declarations, rulesets and so
 on). As a result your CSS becomes much smaller.
}
\examples{
# Merge two rules
csso(".foo { color: #ff0000; }\\n.bar { color: rgba(255, 0, 0, 1); }")

# With a file
path <- system.file("testfiles/example.css", package = "jstools")
csso_file(input = path)

}
