% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jshint.R
\name{jshint}
\alias{jshint}
\alias{jshint_file}
\alias{jshint_options}
\title{JSHint - code validation}
\usage{
jshint_file(input, options = jshint_options())

jshint(code, options = jshint_options())

jshint_options(undef = TRUE, unused = "vars", browser = TRUE,
  jquery = FALSE, devel = FALSE, ...)
}
\arguments{
\item{input}{Path to JavaScript file(s).}

\item{options}{Options for JSHint, see \url{https://jshint.com/docs/options}.}

\item{code}{Character vector where each element represent a line of JavaScript code.}

\item{undef}{This option prohibits the use of explicitly undeclared variables.
This option is very useful for spotting leaking and mistyped variables.}

\item{unused}{This option warns when you define and never use your variables.
It is very useful for general code cleanup, especially when used in addition to undef.}

\item{browser}{This option defines globals exposed by modern browsers: all the way from
good old document and navigator to the HTML5 FileReader and other new developments
in the browser world.}

\item{jquery}{This option defines globals exposed by the jQuery JavaScript library.}

\item{devel}{This option defines globals that are usually used for logging poor-man's
debugging: console, alert, etc. It is usually a good idea to not ship them in production
because, for example, console.log breaks in legacy versions of Internet Explorer.}

\item{...}{Other options to use, see \url{https://jshint.com/docs/options} for details.}
}
\value{
a \code{list} with the diagnosis of the JavaScript code
}
\description{
JSHint is a tool that helps to detect
 errors and potential problems in your JavaScript code.
}
\examples{

# Check code from a character vector
jshint(c("var a = 2;", "var foo = 1;", "b = 3", "a + c"))


# From a file (can be several)
path <- system.file("testfiles/example.js", package = "jstools")
jshint_file(input = path)
}
