% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crass.R
\name{crass}
\alias{crass}
\alias{crass_file}
\title{crass

A CSS minification, pretty printing, and general utility library written in JS.}
\usage{
crass_file(input, pretty = FALSE, output = NULL)

crass(code, pretty = FALSE)
}
\arguments{
\item{input}{Path to a CSS file.}

\item{pretty}{Pretty print result? Otherwise result is minified.}

\item{output}{Path where to write optimized code.}

\item{code}{Character vector where each element represent a line of CSS code.}
}
\value{
a \code{character}.
}
\description{
crass

A CSS minification, pretty printing, and general utility library written in JS.
}
\examples{

crass("b {font-weight: bold;}")

path <- system.file("testfiles/example.css", package = "jstools")
crass_file(input = path)
}
