% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettier.R
\name{prettier}
\alias{prettier}
\alias{prettier_file}
\alias{prettier_options}
\title{Prettier}
\usage{
prettier_file(input, options = prettier_options(), output = NULL)

prettier(code, options = prettier_options())

prettier_options(...)
}
\arguments{
\item{input}{Path to JavaScript file(s).}

\item{options}{Options for terser, see \url{https://prettier.io/docs/en/options.html }.}

\item{output}{Path where to write optimized code.}

\item{code}{Character vector where each element represent a line of JavaScript code.}

\item{...}{Other options to use, see \url{https://prettier.io/docs/en/options.html} for details.}
}
\value{
a \code{character}.
}
\description{
JavaScript parser, mangler and compressor toolkit for ES6+.
}
\examples{


prettier("foo(reallyLongArg(), omgSoManyParameters(),
         IShouldRefactorThis(), isThereSeriouslyAnotherOne());")

path <- system.file("testfiles/htmlwidgets.js", package = "jstools")
prettier_file(input = path)
}
