% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terser.R
\name{terser}
\alias{terser}
\alias{terser_file}
\alias{terser_options}
\title{Terser - JavaScript compressor}
\usage{
terser_file(input, options = terser_options(), output = NULL)

terser(code, options = terser_options())

terser_options(...)
}
\arguments{
\item{input}{Path to one or more JavaScript files.}

\item{options}{Options for terser, see \url{https://terser.org/docs/api-reference}.}

\item{output}{Path where to write optimized code.}

\item{code}{Character vector where each element represent a line of JavaScript code,
or a list of character vectors.}

\item{...}{Other options to use, see \url{https://terser.org/docs/api-reference} for details.}
}
\value{
a \code{list}.
}
\description{
JavaScript parser, mangler and compressor toolkit for ES6+.
}
\examples{

terser("function add(first, second) { return first + second; }")

# Mimic several files
terser(list(
  "file1.js" = "function add(first, second) { return first + second; }",
  "file2.js" = "add(1 + 2, 3 + 4);"
))
# equivalent to:
#> terser_file(c("file1.js", "file2.js"))

# With a file
path <- system.file("testfiles/htmlwidgets.js", package = "jstools")
terser_file(input = path)

}
